#15 Restructured frontend layer
Package "frontend" was renamed to "rest" because is does't contain a real front-end. This is more descriptive. Also, everything thats related to setting up Spring for the REST API was moved to a new sub-package "rest.conf". The "rest" package should mainly contain controllers.
