+++ server/src/main/java/com/xinra/reviewcommunity/rest/conf/FrontendConfiguration.java
@@ -1,4 +1,4 @@
-package com.xinra.reviewcommunity.frontend;
+package com.xinra.reviewcommunity.rest.conf;
 
 import com.xinra.reviewcommunity.MultiMarketMode;
 import org.springframework.beans.factory.annotation.Autowired;
