+++ server/src/main/java/com/xinra/reviewcommunity/rest/conf/RestAuthenticationFailureHandler.java
@@ -1,4 +1,4 @@
-package com.xinra.reviewcommunity.frontend;
+package com.xinra.reviewcommunity.rest.conf;
 
 import java.io.IOException;
 import java.io.PrintWriter;
