+++ server/src/main/java/com/xinra/reviewcommunity/rest/conf/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains everything related to setting up Spring for the REST API.
+ */
+package com.xinra.reviewcommunity.rest.conf;
